Correções de Erros

+++ database/migrations/2025_04_28_193800_create_editor_livro_autor.php
@@ -11,19 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('editor_livro_autor', function (Blueprint $table) {
-            $table->increments('id');
-            $table->foreignId('autor_id')
-                  ->constrained('autors')
-                  ->onDelete('cascade');
-            $table->foreignId('editora_id')
-                  ->constrained('editoras')
-                  ->onDelete('cascade');
-            $table->foreignId('livro_id')
-                  ->constrained('livros')
-                  ->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('editor_livro_autor')) {
+            Schema::create('editor_livro_autor', function (Blueprint $table) {
+                $table->foreignId('autor_id')
+                      ->constrained('autores')
+                      ->onDelete('cascade');
+                $table->foreignId('editora_id')
+                      ->constrained('editoras')
+                      ->onDelete('cascade');
+                $table->foreignId('livro_id')
+                      ->constrained('livros')
+                      ->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
